requirements.txt
Django==2.0.13
djangorestframework==3.9.1
pytz==2018.9
